+++ src/app/services/current-permissions.service.ts
@@ -19,6 +19,11 @@ export enum EGlobalPermissions {
     head = 'head'
 }
 
+enum LocalStorageKeys {
+    permissions_update = 'permissions-update',
+    permissions = 'permissions'
+}
+
 @Injectable()
 export class PermissionsService extends UserService {
 
@@ -40,53 +45,6 @@ export class PermissionsService extends UserService {
         throw new Error(`You are not logged in!`);
     }
 
-    public async isProjectEditor(projectId: number) {
-        return this.hasProjectPermissions(projectId, [EGlobalPermissions.manager],
-            [ELocalPermissions.admin, ELocalPermissions.manager, ELocalPermissions.engineer]);
-    }
-
-    public isGlobalAdmin(): boolean {
-        return !!this.currentUser().admin;
-    }
-
-    public isGlobalManager(): boolean {
-        return !!this.currentUser().manager;
-    }
-
-    public isLocalAdmin(): Promise<boolean> {
-        return this.hasLocalPermissions([ELocalPermissions.admin]);
-    }
-
-    public isLocalManager(): Promise<boolean> {
-        return this.hasLocalPermissions([ELocalPermissions.manager]);
-    }
-
-    public isLocalEngineer(): Promise<boolean> {
-        return this.hasLocalPermissions([ELocalPermissions.manager, ELocalPermissions.engineer]);
-    }
-
-    public isLocalViewer(): Promise<boolean> {
-        return this.hasLocalPermissions([ELocalPermissions.admin, ELocalPermissions.manager,
-        ELocalPermissions.engineer, ELocalPermissions.viewer]);
-    }
-
-    public isProjectAdmin(projectId: number): Promise<boolean> {
-        return this.hasProjectLocalPermissions(projectId, [ELocalPermissions.admin]);
-    }
-
-    public isProjectManager(projectId: number): Promise<boolean> {
-        return this.hasProjectLocalPermissions(projectId, [ELocalPermissions.manager]);
-    }
-
-    public isProjectEngineer(projectId: number): Promise<boolean> {
-        return this.hasProjectLocalPermissions(projectId, [ELocalPermissions.manager, ELocalPermissions.engineer]);
-    }
-
-    public isProjectViewer(projectId: number): Promise<boolean> {
-        return this.hasProjectLocalPermissions(projectId,
-            [ELocalPermissions.admin, ELocalPermissions.manager, ELocalPermissions.engineer, ELocalPermissions.viewer]);
-    }
-
     private hasGlobalPermissions(anyOf: EGlobalPermissions[]) {
         if (!anyOf) {
             return false;
@@ -95,12 +53,13 @@ export class PermissionsService extends UserService {
     }
 
     private async hasProjectLocalPermissions(projectId: number, anyOf: ELocalPermissions[]): Promise<boolean> {
-        const permissions: LocalPermissions[] = await this.getProjectUsers(projectId).toPromise();
+        let permissions: LocalPermissions[] = await this.getLocalPermissions();
+        permissions = permissions.filter(permission => permission.project_id === Number.parseInt(`${projectId}`));
         return this.searchForLocalPermissions(anyOf, permissions);
     }
 
     private async hasLocalPermissions(anyOf: ELocalPermissions[]): Promise<boolean> {
-        const permissions: LocalPermissions[] = await this.getUserProjects(this.currentUser().id);
+        const permissions: LocalPermissions[] = await this.getLocalPermissions();
         return this.searchForLocalPermissions(anyOf, permissions);
     }
 
@@ -111,12 +70,29 @@ export class PermissionsService extends UserService {
         return permissions.find(permission => {
             let result = false;
             anyOf.forEach(name => {
-                if (permission[name] === 1 && this.currentUser().id === permission.user_id) {
+                if (permission[name] === 1) {
                     result = true;
                     return;
                 }
             });
             return result;
         }) !== undefined;
+    }
+
+    private async getLocalPermissions(): Promise<LocalPermissions[]> {
+        if (localStorage.getItem(LocalStorageKeys.permissions_update) !== null) {
+            const update: Date = new Date(localStorage.getItem(LocalStorageKeys.permissions_update));
+            const expiration: Date = new Date();
+            expiration.setMinutes(expiration.getMinutes() - 5);
+
+            if (update > expiration) {
+                return JSON.parse(localStorage.getItem(LocalStorageKeys.permissions));
+            }
+        }
+
+        const newPermissions: LocalPermissions[] = await this.getUserProjects(this.currentUser().id);
+        localStorage.setItem(LocalStorageKeys.permissions, JSON.stringify(newPermissions));
+        localStorage.setItem(LocalStorageKeys.permissions_update, new Date().toString());
+        return JSON.parse(localStorage.getItem(LocalStorageKeys.permissions));
     }
 }